project/requirements/base.txt
# Base
django
django-macros-url
django-template-names
django-split-settings

## Database drivers. Uncomment anyone is used.
##django-pyodbc-azure
##MySQL-python
##mysqlclient
##psycopg2

## cache
##django-redis
##python-memcached
##django-clear-cache

## Project custom app
##django-datatables-view
django-crispy-forms
django-braces>=1.8.1
django-admin-bootstrapped>=2.5.6
django-authtools>=1.4.0
easy-thumbnails>=2.3
